@@ -39,46 +34,35 @@ ext.modauthor = "${modAuthor}"
 ext.nextmcversion = getNextVersion(project.mcversion)
 ext.puzzlesversion = "${puzzlesVersion.substring(1, puzzlesVersion.indexOf("-"))}"
 
-java.toolchain.languageVersion = JavaLanguageVersion.of(8)
+java.toolchain.languageVersion = JavaLanguageVersion.of(16)
 println('Java: ' + System.getProperty('java.version') + ' JVM: ' + System.getProperty('java.vm.version') + '(' + System.getProperty('java.vendor') + ') Arch: ' + System.getProperty('os.arch'))
 
 minecraft {
-
     mappings channel: "${mappingsChannel}", version: "${mappingsVersion}"
-
     runs {
-
         client {
-
             // https://forums.minecraftforge.net/topic/82228-1152-3110-intellij-and-gradlew-forge-hotswap-and-dcevm-tutorial/
             if (project.hasProperty('hotswapAgent') && project.findProperty('enableHotswap').toBoolean()) {
-
                 jvmArgs '-ea', '-Xms1G', '-Xmx4G', '-agentlib:jdwp=transport=dt_socket,server=y,suspend=n,address=5010', '-XXaltjvm=dcevm', '-XX:+UnlockExperimentalVMOptions', "-javaagent:${hotswapAgent}"
             } else {
-
                 jvmArgs '-ea', '-Xms1G', '-Xmx4G'
             }
-
+            property 'fml.earlyprogresswindow', 'false'
+            // mixin properties
             args '--mixin.config', "META-INF/${project.modid}.mixins.json".toString()
-            properties 'fml.earlyprogresswindow': 'false', 'mixin.env.remapRefMap': 'true', 'mixin.env.refMapRemappingFile': "${project.projectDir}/build/createSrgToMcp/output.srg".toString()
+            properties 'mixin.env.remapRefMap': 'true', 'mixin.env.refMapRemappingFile': "${project.projectDir}/build/createSrgToMcp/output.srg".toString()
         }
-
         server {
-
             jvmArgs '-ea', '-Xms1G', '-Xmx4G'
-            args 'nogui', '--mixin.config', "META-INF/${project.modid}.mixins.json".toString()
+            arg 'nogui'
+            // mixin properties
+            args '--mixin.config', "META-INF/${project.modid}.mixins.json".toString()
             properties 'mixin.env.remapRefMap': 'true', 'mixin.env.refMapRemappingFile': "${project.projectDir}/build/createSrgToMcp/output.srg".toString()
         }
     }
 }
 
-//configurations {
-//
-//    shade
-//}
-
 repositories {
-
     mavenLocal()
     maven { url = "https://cursemaven.com" }
     maven {
