@@ -1,50 +1,48 @@
 package fuzs.stylisheffects.client.handler;
 
+import com.mojang.blaze3d.vertex.PoseStack;
 import fuzs.stylisheffects.StylishEffects;
 import fuzs.stylisheffects.client.gui.effects.AbstractEffectRenderer;
 import fuzs.stylisheffects.client.gui.effects.CompactEffectRenderer;
 import fuzs.stylisheffects.client.gui.effects.VanillaEffectRenderer;
 import fuzs.stylisheffects.config.ClientConfig;
-import fuzs.stylisheffects.mixin.client.accessor.DisplayEffectsScreenAccessor;
+import fuzs.stylisheffects.mixin.client.accessor.EffectRenderingInventoryScreenAccessor;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.DisplayEffectsScreen;
-import net.minecraft.client.gui.recipebook.IRecipeShownListener;
-import net.minecraft.client.gui.screen.Screen;
-import net.minecraft.client.gui.screen.inventory.ContainerScreen;
-import net.minecraft.client.gui.screen.inventory.InventoryScreen;
+import net.minecraft.client.gui.screens.Screen;
+import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
+import net.minecraft.client.gui.screens.inventory.EffectRenderingInventoryScreen;
+import net.minecraft.client.gui.screens.inventory.InventoryScreen;
+import net.minecraft.client.gui.screens.recipebook.RecipeUpdateListener;
 import net.minecraftforge.client.event.GuiContainerEvent;
 import net.minecraftforge.client.event.GuiScreenEvent;
-import net.minecraftforge.client.event.RenderGameOverlayEvent;
-import net.minecraftforge.client.event.RenderGameOverlayEvent.ElementType;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
 import javax.annotation.Nullable;
 
 public class EffectScreenHandler {
-    @Nullable
-    public static AbstractEffectRenderer inventoryRenderer;
-    @Nullable
-    public static AbstractEffectRenderer hudRenderer;
+    public static final EffectScreenHandler INSTANCE = new EffectScreenHandler();
 
-    public static void createEffectRenderers() {
-        inventoryRenderer = createEffectRenderer(StylishEffects.CONFIG.client().inventoryRenderer().rendererType, AbstractEffectRenderer.EffectRendererType.INVENTORY);
-        hudRenderer = createEffectRenderer(StylishEffects.CONFIG.client().hudRenderer().rendererType, AbstractEffectRenderer.EffectRendererType.HUD);
+    @Nullable
+    public AbstractEffectRenderer inventoryRenderer;
+    @Nullable
+    public AbstractEffectRenderer hudRenderer;
+
+    private EffectScreenHandler() {
+
+    }
+
+    public void createEffectRenderers() {
+        this.inventoryRenderer = createEffectRenderer(StylishEffects.CONFIG.client().inventoryRenderer().rendererType, AbstractEffectRenderer.EffectRendererType.INVENTORY);
+        this.hudRenderer = createEffectRenderer(StylishEffects.CONFIG.client().hudRenderer().rendererType, AbstractEffectRenderer.EffectRendererType.HUD);
     }
 
     @Nullable
     private static AbstractEffectRenderer createEffectRenderer(ClientConfig.EffectRenderer rendererType, AbstractEffectRenderer.EffectRendererType effectRendererType) {
-        switch (rendererType) {
-            case VANILLA:
-                return new VanillaEffectRenderer(effectRendererType);
-            case COMPACT:
-                return new CompactEffectRenderer(effectRendererType);
-        }
-        return null;
-    }
-
-    @SubscribeEvent
-    public void onRenderGameOverlayPre(final RenderGameOverlayEvent.Pre evt) {
-        if (evt.getType() == ElementType.POTION_ICONS) evt.setCanceled(true);
+        return switch (rendererType) {
+            case VANILLA -> new VanillaEffectRenderer(effectRendererType);
+            case COMPACT -> new CompactEffectRenderer(effectRendererType);
+            default -> null;
+        };
     }
 
     @SubscribeEvent
@@ -54,32 +52,31 @@ public class EffectScreenHandler {
 
     @SubscribeEvent
     public void onInitGuiPost(final GuiScreenEvent.InitGuiEvent.Post evt) {
-        if (evt.getGui() instanceof DisplayEffectsScreen) {
+        if (evt.getGui() instanceof EffectRenderingInventoryScreen) {
             // disable vanilla rendering in creative mode inventory, survival inventory has to be disabled separately
             // this is not needed by us, we just check before rendering as survival inventory does
-            ((DisplayEffectsScreenAccessor) evt.getGui()).setDoRenderEffects(false);
+            ((EffectRenderingInventoryScreenAccessor) evt.getGui()).setDoRenderEffects(false);
         }
     }
 
     @SubscribeEvent
     public void onDrawBackground(final GuiContainerEvent.DrawBackground evt) {
         if (evt.getGuiContainer() instanceof InventoryScreen) {
-            ((DisplayEffectsScreenAccessor) evt.getGuiContainer()).setDoRenderEffects(false);
+            ((EffectRenderingInventoryScreenAccessor) evt.getGuiContainer()).setDoRenderEffects(false);
         }
     }
 
-    @SubscribeEvent
-    public void onRenderGameOverlayText(final RenderGameOverlayEvent.Text evt) {
+    public void onRenderGameOverlayText(PoseStack poseStack, int screenWidth, int screenHeight) {
         // use this event so potion icons are drawn behind debug menu as in vanilla
         // field may get changed during config reload from different thread
-        final AbstractEffectRenderer hudRenderer = EffectScreenHandler.hudRenderer;
+        final AbstractEffectRenderer hudRenderer = this.hudRenderer;
         if (hudRenderer != null) {
             final Minecraft minecraft = Minecraft.getInstance();
             if (!minecraft.player.getActiveEffects().isEmpty()) {
                 final ClientConfig.ScreenSide screenSide = StylishEffects.CONFIG.client().hudRenderer().screenSide;
-                hudRenderer.setScreenDimensions(minecraft.gui, evt.getWindow().getGuiScaledWidth(), evt.getWindow().getGuiScaledHeight(), screenSide.right() ? evt.getWindow().getGuiScaledWidth() : 0, 0, screenSide);
+                hudRenderer.setScreenDimensions(minecraft.gui, screenWidth, screenHeight, screenSide.right() ? screenWidth : 0, 0, screenSide);
                 hudRenderer.setActiveEffects(minecraft.player.getActiveEffects());
-                hudRenderer.renderEffects(evt.getMatrixStack(), minecraft);
+                hudRenderer.renderEffects(poseStack, minecraft);
             }
         }
     }
@@ -87,9 +84,9 @@ public class EffectScreenHandler {
     @SubscribeEvent
     public void onDrawScreenPost(final GuiScreenEvent.DrawScreenEvent.Post evt) {
         // field may get changed during config reload from different thread
-        final AbstractEffectRenderer inventoryRenderer = EffectScreenHandler.inventoryRenderer;
-        if (inventoryRenderer != null && supportsEffectsDisplay(evt.getGui())) {
-            ContainerScreen<?> screen = (ContainerScreen<?>) evt.getGui();
+        final AbstractEffectRenderer inventoryRenderer = this.inventoryRenderer;
+        if (inventoryRenderer != null && this.supportsEffectsDisplay(evt.getGui())) {
+            AbstractContainerScreen<?> screen = (AbstractContainerScreen<?>) evt.getGui();
             final Minecraft minecraft = screen.getMinecraft();
             if (!minecraft.player.getActiveEffects().isEmpty()) {
                 final ClientConfig.ScreenSide screenSide = StylishEffects.CONFIG.client().inventoryRenderer().screenSide;
@@ -101,13 +98,13 @@ public class EffectScreenHandler {
         }
     }
 
-    public static boolean supportsEffectsDisplay(Screen screen) {
-        if (screen instanceof DisplayEffectsScreen) {
+    public boolean supportsEffectsDisplay(Screen screen) {
+        if (screen instanceof EffectRenderingInventoryScreen) {
             return true;
         }
-        if (StylishEffects.CONFIG.client().inventoryRenderer().effectsEverywhere && screen instanceof ContainerScreen) {
-            if (screen instanceof IRecipeShownListener) {
-                if (((IRecipeShownListener) screen).getRecipeBookComponent().isVisible()) {
+        if (StylishEffects.CONFIG.client().inventoryRenderer().effectsEverywhere && screen instanceof AbstractContainerScreen) {
+            if (screen instanceof RecipeUpdateListener) {
+                if (((RecipeUpdateListener) screen).getRecipeBookComponent().isVisible()) {
                     return StylishEffects.CONFIG.client().inventoryRenderer().screenSide == ClientConfig.ScreenSide.RIGHT;
                 }
             }
