@@ -1,9 +1,9 @@
 package fuzs.stylisheffects.client.gui.effects;
 
-import net.minecraft.client.renderer.Rectangle2d;
+import net.minecraft.client.renderer.Rect2i;
 
 import java.util.List;
 
 public interface IHasRenderAreas {
-    List<Rectangle2d> getRenderAreas();
+    List<Rect2i> getRenderAreas();
 }
