@@ -1,17 +1,19 @@
 package fuzs.stylisheffects.client.gui.effects;
 
 import com.google.common.collect.Lists;
-import com.mojang.blaze3d.matrix.MatrixStack;
 import com.mojang.blaze3d.systems.RenderSystem;
+import com.mojang.blaze3d.vertex.PoseStack;
 import fuzs.stylisheffects.StylishEffects;
 import fuzs.stylisheffects.config.ClientConfig;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.AbstractGui;
-import net.minecraft.client.renderer.texture.PotionSpriteUploader;
+import net.minecraft.client.gui.GuiComponent;
+import net.minecraft.client.renderer.GameRenderer;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
-import net.minecraft.potion.EffectInstance;
-import net.minecraft.util.text.IFormattableTextComponent;
-import net.minecraft.util.text.TranslationTextComponent;
+import net.minecraft.client.resources.MobEffectTextureManager;
+import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraftforge.client.RenderProperties;
 import org.apache.commons.lang3.tuple.Pair;
 
 import java.util.List;
@@ -37,10 +39,10 @@ public class VanillaEffectRenderer extends AbstractEffectRenderer {
     }
 
     @Override
-    public List<Pair<EffectInstance, int[]>> getEffectPositions(List<EffectInstance> activeEffects) {
+    public List<Pair<MobEffectInstance, int[]>> getEffectPositions(List<MobEffectInstance> activeEffects) {
         int counter = 0;
-        List<Pair<EffectInstance, int[]>> effectToPos = Lists.newArrayList();
-        for (EffectInstance effect : activeEffects) {
+        List<Pair<MobEffectInstance, int[]>> effectToPos = Lists.newArrayList();
+        for (MobEffectInstance effect : activeEffects) {
             int posX = counter % this.getMaxColumns();
             int posY = counter / this.getMaxColumns();
             counter++;
@@ -52,30 +54,31 @@ public class VanillaEffectRenderer extends AbstractEffectRenderer {
     }
 
     @Override
-    public void renderWidget(MatrixStack matrixStack, int posX, int posY, Minecraft minecraft, EffectInstance effectinstance) {
+    public void renderWidget(PoseStack matrixStack, int posX, int posY, Minecraft minecraft, MobEffectInstance effectinstance) {
         RenderSystem.enableBlend();
-        minecraft.getTextureManager().bind(EFFECT_BACKGROUND);
-        RenderSystem.color4f(1.0F, 1.0F, 1.0F, this.config().widgetAlpha);
-        AbstractGui.blit(matrixStack, posX, posY, 0, StylishEffects.CONFIG.client().vanillaWidget().ambientBorder && effectinstance.isAmbient() ? this.getHeight() : 0, this.getWidth(), this.getHeight(), 256, 256);
+        RenderSystem.setShader(GameRenderer::getPositionTexShader);
+        RenderSystem.setShaderTexture(0, EFFECT_BACKGROUND);
+        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, this.config().widgetAlpha);
+        GuiComponent.blit(matrixStack, posX, posY, 0, StylishEffects.CONFIG.client().vanillaWidget().ambientBorder && effectinstance.isAmbient() ? this.getHeight() : 0, this.getWidth(), this.getHeight(), 256, 256);
         this.drawEffectSprite(matrixStack, posX, posY, minecraft, effectinstance);
         this.drawCustomEffect(matrixStack, posX, posY, effectinstance);
         this.drawEffectText(matrixStack, posX, posY, minecraft, effectinstance);
     }
 
-    private void drawEffectSprite(MatrixStack matrixStack, int posX, int posY, Minecraft minecraft, EffectInstance effectinstance) {
-        PotionSpriteUploader potionspriteuploader = minecraft.getMobEffectTextures();
+    private void drawEffectSprite(PoseStack matrixStack, int posX, int posY, Minecraft minecraft, MobEffectInstance effectinstance) {
+        MobEffectTextureManager potionspriteuploader = minecraft.getMobEffectTextures();
         TextureAtlasSprite textureatlassprite = potionspriteuploader.get(effectinstance.getEffect());
-        minecraft.getTextureManager().bind(textureatlassprite.atlas().location());
+        RenderSystem.setShaderTexture(0, textureatlassprite.atlas().location());
         final float blinkingAlpha = StylishEffects.CONFIG.client().vanillaWidget().blinkingAlpha ? this.getBlinkingAlpha(effectinstance) : 1.0F;
-        RenderSystem.color4f(1.0F, 1.0F, 1.0F, blinkingAlpha * this.config().widgetAlpha);
-        AbstractGui.blit(matrixStack, posX + 6, posY + 7, 0, 18, 18, textureatlassprite);
+        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, blinkingAlpha * this.config().widgetAlpha);
+        GuiComponent.blit(matrixStack, posX + 6, posY + 7, 0, 18, 18, textureatlassprite);
     }
 
-    private void drawEffectText(MatrixStack matrixStack, int posX, int posY, Minecraft minecraft, EffectInstance effectinstance) {
-        if (effectinstance.shouldRenderInvText()) {
-            IFormattableTextComponent component = new TranslationTextComponent(effectinstance.getEffect().getDescriptionId());
+    private void drawEffectText(PoseStack matrixStack, int posX, int posY, Minecraft minecraft, MobEffectInstance effectinstance) {
+        if (RenderProperties.getEffectRenderer(effectinstance).shouldRenderInvText(effectinstance)) {
+            MutableComponent component = new TranslatableComponent(effectinstance.getEffect().getDescriptionId());
             if (effectinstance.getAmplifier() >= 1 && effectinstance.getAmplifier() <= 9) {
-                component.append(" ").append(new TranslationTextComponent("enchantment.level." + (effectinstance.getAmplifier() + 1)));
+                component.append(" ").append(new TranslatableComponent("enchantment.level." + (effectinstance.getAmplifier() + 1)));
             }
             int nameColor = ColorUtil.getEffectColor(StylishEffects.CONFIG.client().vanillaWidget().nameColor, effectinstance);
             minecraft.font.drawShadow(matrixStack, component, posX + 10 + 18, posY + 7 + (!StylishEffects.CONFIG.client().vanillaWidget().ambientDuration && effectinstance.isAmbient() ? 4 : 0), (int) (this.config().widgetAlpha * 255.0F) << 24 | nameColor);
