@@ -4,26 +4,26 @@ import fuzs.stylisheffects.client.gui.effects.AbstractEffectRenderer;
 import fuzs.stylisheffects.client.handler.EffectScreenHandler;
 import mezz.jei.api.gui.handlers.IGuiContainerHandler;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.screen.inventory.ContainerScreen;
-import net.minecraft.client.renderer.Rectangle2d;
-import net.minecraft.inventory.container.Container;
-import net.minecraft.potion.EffectInstance;
+import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
+import net.minecraft.client.renderer.Rect2i;
+import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraft.world.inventory.AbstractContainerMenu;
 
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
-class EffectRendererGuiHandler<T extends Container> implements IGuiContainerHandler<ContainerScreen<T>> {
+class EffectRendererGuiHandler<T extends AbstractContainerMenu> implements IGuiContainerHandler<AbstractContainerScreen<T>> {
 
     @Override
-    public List<Rectangle2d> getGuiExtraAreas(ContainerScreen<T> screen) {
+    public List<Rect2i> getGuiExtraAreas(AbstractContainerScreen<T> screen) {
         // field may get changed during config reload from different thread
-        final AbstractEffectRenderer inventoryRenderer = EffectScreenHandler.inventoryRenderer;
-        if (inventoryRenderer != null && EffectScreenHandler.supportsEffectsDisplay(screen)) {
+        final AbstractEffectRenderer inventoryRenderer = EffectScreenHandler.INSTANCE.inventoryRenderer;
+        if (inventoryRenderer != null && EffectScreenHandler.INSTANCE.supportsEffectsDisplay(screen)) {
             Minecraft minecraft = screen.getMinecraft();
-            Collection<EffectInstance> activePotionEffects = minecraft.player.getActiveEffects();
+            Collection<MobEffectInstance> activePotionEffects = minecraft.player.getActiveEffects();
             if (!activePotionEffects.isEmpty()) {
-                return EffectScreenHandler.inventoryRenderer.getRenderAreas();
+                return EffectScreenHandler.INSTANCE.inventoryRenderer.getRenderAreas();
             }
         }
         return Collections.emptyList();
